added cocoa type, fixed language-crashing conflict
This fixes a conflict in the "func" mode that was making the Swift language error out during highlighting some code samples. Also adds a mode that matches UI-, NS- and CG- prefixed types (these are the Cocoa/Foundation types in use on OS X and iOS) for stronger Swift matching.